- added js code to verify contracts at deployment
- added formatting / linting for sol to the ci
- set env var FRESH_DEPLOYMENT to true to trigger a "fresh" deployment (all contracts will be deployed and respective files updated with the new contract addresses)

+++ contracts/base/DiamondStorageBase.sol
@@ -14,25 +14,24 @@ contract DiamondStorageBase is EternalStorage {
         // and maps the selectors to the position in the slot.
         // func selector => address facet, uint64 slotsIndex, uint64 slotIndex
         mapping(bytes4 => bytes32) facets;
-
         // array of slots of function selectors.
         // each slot holds 8 function selectors.
-        mapping(uint => bytes32) selectorSlots;
-
+        mapping(uint256 => bytes32) selectorSlots;
         // uint128 numSelectorsInSlot, uint128 selectorSlotsLength
         // selectorSlotsLength is the number of 32-byte slots in selectorSlots.
         // selectorSlotLength is the number of selectors in the last slot of
         // selectorSlots.
-        uint selectorSlotsLength;
-
+        uint256 selectorSlotsLength;
         // tracking initialization state
         // we use this to know whether a call to diamondCut() is part of the initial
         // construction or a later "upgrade" call
         bool initialized;
     }
 
-    function diamondStorage() internal pure returns(DiamondStorage storage ds) {
+    function diamondStorage() internal pure returns (DiamondStorage storage ds) {
         // ds_slot = keccak256("diamond.standard.diamond.storage");
-        assembly { ds.slot := 0xc8fcad8db84d3cc18b4c41d551ea0ee66dd599cde068d998e57d5e09332c131c }
+        assembly {
+            ds.slot := 0xc8fcad8db84d3cc18b4c41d551ea0ee66dd599cde068d998e57d5e09332c131c
+        }
     }
 }
